类型转换警告处理

+++ src/Camera.cpp
@@ -28,7 +28,7 @@ int Camera::getOffsetX()
     {
         tx = WORLD_RIGHT - GAME_WIDTH;
     }
-    return tx;
+    return static_cast<int>(tx);
 } 
 
 int Camera::getOffsetY()
